Conflicts:
	src/Agence.java
	src/Lieu.java
	src/LieuFormation.java

+++ src/Lieu.java
@@ -15,9 +15,9 @@ public abstract class Lieu {
     protected String name;
     protected String id;
     
-    public Lieu (String id, String name, int codepostale, int longitude, int latitude)
+    public Lieu (String id , String name, int codepostale, float longitude, float latitude)
     {
-        this.id = id;
+    	this.id = id;
         this.name = name;
         this.codepostale = codepostale;
         this.longitude = longitude;
